Theme admin_future for users:user_2step

+++ src/modules/users/admin/edit_2step.php
@@ -46,36 +46,35 @@ if (!$allow) {
 }
 
 if ($admin_info['admin_id'] == $userid and $admin_info['safemode'] == 1) {
-    $xtpl = new XTemplate('user_safemode.tpl', NV_ROOTDIR . '/themes/' . $global_config['module_theme'] . '/modules/' . $module_file);
-    $xtpl->assign('LANG', \NukeViet\Core\Language::$lang_module);
-    $xtpl->assign('SAFEMODE_DEACT', NV_BASE_SITEURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&amp;' . NV_NAME_VARIABLE . '=users&amp;' . NV_OP_VARIABLE . '=editinfo/safeshow');
-    $xtpl->parse('main');
-    $contents = $xtpl->text('main');
+    $tpl = new \NukeViet\Template\NVSmarty();
+    $tpl->setTemplateDir(get_module_tpl_dir('user_safemode.tpl'));
+    $tpl->assign('LANG', $nv_Lang);
+    $tpl->assign('MODULE_NAME', $module_name);
+    $tpl->assign('OP', $op);
+    $contents = $tpl->fetch('user_safemode.tpl');
 
     include NV_ROOTDIR . '/includes/header.php';
     echo nv_admin_theme($contents);
     include NV_ROOTDIR . '/includes/footer.php';
-    exit();
 }
 
 // Thêm vào menutop
 $select_options[NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&amp;' . NV_NAME_VARIABLE . '=' . $module_name . '&amp;' . NV_OP_VARIABLE . '=edit&amp;userid=' . $row['userid']] = $nv_Lang->getModule('edit_title');
 $select_options[NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&amp;' . NV_NAME_VARIABLE . '=' . $module_name . '&amp;' . NV_OP_VARIABLE . '=edit_oauth&amp;userid=' . $row['userid']] = $nv_Lang->getModule('user_openid_mamager');
 
-$xtpl = new XTemplate('user_2step.tpl', NV_ROOTDIR . '/themes/' . $global_config['module_theme'] . '/modules/' . $module_file);
-$xtpl->assign('LANG', \NukeViet\Core\Language::$lang_module);
-$xtpl->assign('FORM_ACTION', NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&amp;' . NV_NAME_VARIABLE . '=' . $module_name . '&amp;' . NV_OP_VARIABLE . '=' . $op . '&amp;userid=' . $row['userid']);
+$codes = [];
 
-if (empty($row['active2step'])) {
-    $xtpl->parse('turnoff');
-    $contents = $xtpl->text('turnoff');
-} else {
-    if (!empty($global_config['two_step_verification'])) {
-        $xtpl->parse('main.turnoff_info');
-    }
-
+if (!empty($row['active2step'])) {
     // Tắt xác thực hai bước
     if ($nv_Request->isset_request('turnoff2step', 'post')) {
+        if (!defined('NV_IS_AJAX')) {
+            exit('Wrong URL');
+        }
+
+        if (!csrf_check($nv_Request->get_string('checkss', 'post'), $csrf_key)) {
+            nv_jsonOutput(['status' => 'error', 'mess' => $nv_Lang->getGlobal('error_checkss')]);
+        }
+
         $db->query('DELETE FROM ' . NV_MOD_TABLE . '_backupcodes WHERE userid=' . $row['userid']);
         $db->query('DELETE FROM ' . NV_MOD_TABLE . '_passkey WHERE userid=' . $row['userid'] . ' AND enable_login=0');
         $db->query('UPDATE ' . NV_MOD_TABLE . " SET active2step=0, secretkey='', last_update=" . NV_CURRENTTIME . ' WHERE userid=' . $row['userid']);
@@ -109,13 +108,22 @@ if (empty($row['active2step'])) {
 
         nv_insert_logs(NV_LANG_DATA, $module_name, 'log_turnoff_user2step', 'userid ' . $row['userid'], $admin_info['userid']);
         $nv_Cache->delMod($module_name);
-
-        header('Location: ' . NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&' . NV_NAME_VARIABLE . '=' . $module_name . '&' . NV_OP_VARIABLE . '=' . $op . '&userid=' . $row['userid']);
-        exit();
+        nv_jsonOutput([
+            'status' => 'OK',
+            'mess' => $nv_Lang->getModule('user_2step_turnoff')
+        ]);
     }
 
     // Tạo lại mã dự phòng
     if ($nv_Request->isset_request('resetbackupcodes', 'post')) {
+        if (!defined('NV_IS_AJAX')) {
+            exit('Wrong URL');
+        }
+
+        if (!csrf_check($nv_Request->get_string('checkss', 'post'), $csrf_key)) {
+            nv_jsonOutput(['status' => 'error', 'mess' => $nv_Lang->getGlobal('error_checkss')]);
+        }
+
         $db->query('DELETE FROM ' . NV_MOD_TABLE . '_backupcodes WHERE userid=' . $row['userid']);
 
         $new_code = [];
@@ -158,22 +166,31 @@ if (empty($row['active2step'])) {
 
         nv_insert_logs(NV_LANG_DATA, $module_name, 'log_reset_user2step_codes', 'userid ' . $row['userid'], $admin_info['userid']);
         $nv_Cache->delMod($module_name);
-        nv_redirect_location(NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&' . NV_NAME_VARIABLE . '=' . $module_name . '&' . NV_OP_VARIABLE . '=' . $op . '&userid=' . $row['userid']);
+        nv_jsonOutput([
+            'status' => 'OK',
+            'mess' => $nv_Lang->getModule('user_2step_reset')
+        ]);
     }
 
     $sql = 'SELECT * FROM ' . NV_MOD_TABLE . '_backupcodes WHERE userid=' . $row['userid'];
     $result = $db->query($sql);
     while ($code = $result->fetch()) {
-        $code['status'] = $nv_Lang->getModule('user_2step_codes_s' . $code['is_used']);
-        $code['time_creat'] = $code['time_creat'] ? nv_datetime_format($code['time_creat'] , 1) : '';
-        $code['time_used'] = $code['time_used'] ? nv_datetime_format($code['time_used'], 1) : '';
-        $xtpl->assign('CODE', $code);
-        $xtpl->parse('main.code');
+        $code['status_label'] = $nv_Lang->getModule('user_2step_codes_s' . $code['is_used']);
+        $codes[] = $code;
     }
-
-    $xtpl->parse('main');
-    $contents = $xtpl->text('main');
 }
+
+$tpl = new \NukeViet\Template\NVSmarty();
+$tpl->setTemplateDir(get_module_tpl_dir('user_2step.tpl'));
+$tpl->registerPlugin('modifier', 'ddatetime', 'nv_datetime_format');
+$tpl->assign('LANG', $nv_Lang);
+$tpl->assign('MODULE_NAME', $module_name);
+$tpl->assign('OP', $op);
+$tpl->assign('CHECKSS', csrf_create($csrf_key));
+$tpl->assign('ROW', $row);
+$tpl->assign('GCONFIG', $global_config);
+$tpl->assign('CODES', $codes);
+$contents = $tpl->fetch('user_2step.tpl');
 
 include NV_ROOTDIR . '/includes/header.php';
 echo nv_admin_theme($contents);
